requirements.txt
numpy
PyQt5==5.15.9
python-telegram-bot==20.5
requests==2.31.0
pandas==2.2.2
python-binance==1.0.16
cryptography==41.0.3
nest_asyncio==1.5.6
aiohttp==3.8.4
uvloop; platform_system!='Windows'
